Added new markers
Changed the css to display in a grid. Looks neater and more organized. Added labels for the new markers added. Basic logic set up for the markers, red for selected and gray for not selected. This is to be used once I get into making routes for directions. Added a basic infoWindow that shows the location name (address soon), and the level.